BaseGameController - all common endpoints for games are now in basic controller. ExampleGameController is provided to show how a new controller can be made. You can also see how Kik is implemented (don't forget to update app.js)
